+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.2.3] - 2022-02-02
+
+### Added
+ - support for CSV based spectra
+ - support for CSV based spectrograph efficiency files
+
+### Changed
+ - updated project dependencies
+
 ## [0.2.2] - 2022-01-26
 
 ### Fixed